+++ disassemblers/ofrak_capstone/ofrak_capstone/components.py
@@ -143,7 +143,7 @@ class CapstoneInstructionRegisterUsageAnalyzer(InstructionRegisterUsageAnalyzer)
         super().__init__(resource_factory, data_service, resource_service)
         self._disassembler_service = disassembler_service
 
-    async def analyze(self, resource: Resource, config: None) -> RegisterUsage:
+    async def analyze(self, resource: Resource, config=None) -> RegisterUsage:
         program_attrs = await resource.analyze(ProgramAttributes)
 
         instruction = await resource.view_as(Instruction)