+++ frontend/src/ofrak/resource.js
@@ -386,6 +386,22 @@ export class Resource {
     throw new NotImplementedError("update_script");
   }
 
+  async get_components() {
+    throw new NotImplementedError("get_components");
+  }
+
+  async get_config_for_component() {
+    throw new NotImplementedError("get_config_for_component");
+  }
+
+  async run_component() {
+    throw new NotImplementedError("run_component");
+  }
+
+  async get_tags_and_num_components() {
+    throw new NotImplementedError("get_tags_and_num_components");
+  }
+
   async add_flush_to_disk_to_script() {
     throw new NotImplementedError("add_flush_to_disk_to_script");
   }