+++ ofrak_core/test_ofrak/unit/test_ofrak_server.py
@@ -10,9 +10,11 @@ from typing import List
 
 from aiohttp.test_utils import TestClient
 
+from ofrak import Analyzer, Unpacker, Modifier, Packer
 from ofrak.core import File
 from ofrak.core.entropy import DataSummaryAnalyzer
 from ofrak.gui.server import AiohttpOFRAKServer, start_server
+from ofrak.model.component_filters import ComponentOrMetaFilter, ComponentTypeFilter
 from ofrak.service.serialization.pjson import (
     PJSONSerializationService,
 )
@@ -67,7 +69,6 @@ def dicts_are_similar(d1, d2, attributes_to_skip=None):
     for key, value in d1.items():
         if key in attributes_to_skip:
             continue
-        # Necessary to compare the unordered list of tags returned from unpack
         if isinstance(value, list) and set(value) != set(d2[key]):
             return False
         elif value != d2[key]:
@@ -181,6 +182,29 @@ async def test_get_children(ofrak_client: TestClient, hello_world_elf):
     children_body = await children_resp.json()
     assert root_id in children_body
     assert len(children_body[root_id]) > 1
+
+
+async def test_get_descendants(ofrak_client: TestClient, hello_world_elf):
+    create_resp = await ofrak_client.post(
+        "/create_root_resource", params={"name": "hello_world_elf"}, data=hello_world_elf
+    )
+    root = await create_resp.json()
+    root_id = root["id"]
+    await ofrak_client.post(f"/{root_id}/unpack")
+    children_resp = await ofrak_client.post(f"/batch/get_children", json=[root_id])
+    children = await children_resp.json()
+    children_ids = [
+        cid_v for child in children[root_id] for cid_k, cid_v, in child.items() if cid_k == "id"
+    ]
+    descendants_resp = await ofrak_client.get(f"/{root_id}/get_descendants")
+    assert descendants_resp.status == 200
+    descendants = await descendants_resp.json()
+    descendant_ids = [
+        did_v for descendant in descendants for did_k, did_v in descendant.items() if did_k == "id"
+    ]
+    for child in children_ids:
+        assert child in descendant_ids
+    assert len(descendants) > 1
 
 
 async def test_get_data_range(ofrak_client: TestClient, hello_world_elf):
@@ -441,11 +465,12 @@ async def test_update_script(ofrak_client: TestClient, hello_world_elf):
         "from ofrak.core import *",
         "",
         "",
-        "async def main(ofrak_context: OFRAKContext):",
+        "async def main(ofrak_context: OFRAKContext, root_resource: Optional[Resource] = None):",
         "",
-        "    root_resource = await ofrak_context.create_root_resource_from_file(",
-        '        "hello_world_elf"',
-        "    )",
+        "    if root_resource is None:",
+        "        root_resource = await ofrak_context.create_root_resource_from_file(",
+        '            "hello_world_elf"',
+        "        )",
         "",
         "    await root_resource.unpack()",
         "",
@@ -561,11 +586,12 @@ async def test_selectable_attr_err(ofrak_client: TestClient, hello_world_elf):
         "from ofrak.core import *",
         "",
         "",
-        "async def main(ofrak_context: OFRAKContext):",
+        "async def main(ofrak_context: OFRAKContext, root_resource: Optional[Resource] = None):",
         "",
-        "    root_resource = await ofrak_context.create_root_resource_from_file(",
-        '        "hello_world_elf"',
-        "    )",
+        "    if root_resource is None:",
+        "        root_resource = await ofrak_context.create_root_resource_from_file(",
+        '            "hello_world_elf"',
+        "        )",
         "",
         "    await root_resource.create_child(",
         "        tags=(GenericBinary,), data_range=Range(0x0, 0x1FF5)",
@@ -690,6 +716,191 @@ async def test_clear_action_queue(ofrak_client: TestClient, hello_world_elf):
         "",
     ]
 
+
+async def test_get_components(ofrak_client: TestClient, hello_world_elf, ofrak_context):
+    create_resp = await ofrak_client.post(
+        "/create_root_resource", params={"name": "hello_world_elf"}, data=hello_world_elf
+    )
+    create_body = await create_resp.json()
+    resource_id = create_body["id"]
+    resp = await ofrak_client.post(
+        f"/{resource_id}/get_components",
+        json={
+            "show_all_components": True,
+            "target_filter": None,
+            "analyzers": True,
+            "modifiers": True,
+            "packers": True,
+            "unpackers": True,
+        },
+    )
+    components = set(await resp.json())
+    expected_components = ofrak_context.component_locator.get_components_matching_filter(
+        ComponentOrMetaFilter(
+            ComponentTypeFilter(Analyzer),
+            ComponentTypeFilter(Unpacker),
+            ComponentTypeFilter(Modifier),
+            ComponentTypeFilter(Packer),
+        )
+    )
+    assert components == {
+        type(comp).__name__ for comp in expected_components if "Angr" not in type(comp).__name__
+    }
+
+
+async def test_get_config(ofrak_client: TestClient, hello_world_elf):
+    create_resp = await ofrak_client.post(
+        "/create_root_resource", params={"name": "hello_world_elf"}, data=hello_world_elf
+    )
+    create_body = await create_resp.json()
+    resource_id = create_body["id"]
+    components_resp = await ofrak_client.post(
+        f"/{resource_id}/get_components",
+        json={
+            "show_all_components": True,
+            "target_filter": None,
+            "analyzers": True,
+            "modifiers": True,
+            "packers": True,
+            "unpackers": True,
+        },
+    )
+    components = await components_resp.json()
+    configs = []
+    for component in components:
+        configs_resp = await ofrak_client.get(
+            f"/{resource_id}/get_config_for_component",
+            params={"component": component},
+        )
+        configs.append(await configs_resp.json())
+    assert len(configs) == len(components)
+    config_resp = await ofrak_client.get(
+        f"/{resource_id}/get_config_for_component",
+        params={"component": "UpdateLinkableSymbolsModifier"},
+    )
+    config = await config_resp.json()
+    assert config == {
+        "name": "UpdateLinkableSymbolsModifierConfig",
+        "type": "ofrak.core.patch_maker.linkable_binary.UpdateLinkableSymbolsModifierConfig",
+        "args": None,
+        "enum": None,
+        "fields": [
+            {
+                "name": "updated_symbols",
+                "type": "typing.List",
+                "args": [
+                    {
+                        "name": None,
+                        "type": "ofrak.core.patch_maker.linkable_symbol.LinkableSymbol",
+                        "args": None,
+                        "fields": [
+                            {
+                                "name": "virtual_address",
+                                "type": "builtins.int",
+                                "args": None,
+                                "fields": None,
+                                "enum": None,
+                                "default": None,
+                            },
+                            {
+                                "name": "name",
+                                "type": "builtins.str",
+                                "args": None,
+                                "fields": None,
+                                "enum": None,
+                                "default": None,
+                            },
+                            {
+                                "name": "symbol_type",
+                                "type": "ofrak_type.symbol_type.LinkableSymbolType",
+                                "args": None,
+                                "fields": None,
+                                "enum": {"FUNC": 0, "RW_DATA": 1, "RO_DATA": 2, "UNDEF": -1},
+                                "default": None,
+                            },
+                            {
+                                "name": "mode",
+                                "type": "ofrak_type.architecture.InstructionSetMode",
+                                "args": None,
+                                "fields": None,
+                                "enum": {"NONE": 0, "THUMB": 1, "VLE": 2},
+                                "default": 0,
+                            },
+                        ],
+                        "enum": None,
+                        "default": None,
+                    }
+                ],
+                "fields": None,
+                "enum": None,
+                "default": None,
+            }
+        ],
+    }
+
+
+async def test_get_tags_and_num_components(ofrak_client: TestClient, hello_world_elf):
+    create_resp = await ofrak_client.post(
+        "/create_root_resource", params={"name": "hello_world_elf"}, data=hello_world_elf
+    )
+    create_body = await create_resp.json()
+    resource_id = create_body["id"]
+    resp = await ofrak_client.post(
+        f"/{resource_id}/get_tags_and_num_components",
+        json={
+            "target": "File",
+            "analyzers": True,
+            "modifiers": True,
+            "packers": True,
+            "unpackers": True,
+        },
+    )
+    resp_body = await resp.json()
+    assert (
+        resp.status == 200
+    )  # The result of the components differs based on the number of components in OFRAK, so checking the exact output will break everytime a component is added.
+
+
+async def test_run_component(ofrak_client: TestClient, hello_world_elf):
+    create_resp = await ofrak_client.post(
+        "/create_root_resource", params={"name": "hello_world_elf"}, data=hello_world_elf
+    )
+    create_body = await create_resp.json()
+    resource_id = create_body["id"]
+    resp = await ofrak_client.post(
+        f"/{resource_id}/run_component",
+        params={"component": "StringFindReplaceModifier"},
+        json=[
+            "ofrak.core.strings.StringFindReplaceConfig",
+            {
+                "allow_overflow": False,
+                "null_terminate": False,
+                "to_find": "ELF",
+                "replace_with": "ORC",
+            },
+        ],
+    )
+    assert resp.status == 200
+    resp_body = await resp.json()
+    expected_list = {
+        "created": [],
+        "modified": [
+            {
+                "id": "00000001",
+                "data_id": "00000001",
+                "parent_id": None,
+                "tags": ["ofrak.core.filesystem.File", "ofrak.core.filesystem.FilesystemEntry"],
+                "attributes": [
+                    [
+                        "ofrak.model.resource_model.Data",
+                        ["ofrak.model.resource_model.Data", {"_offset": 0, "_length": 8181}],
+                    ]
+                ],
+                "caption": "File",
+            }
+        ],
+        "deleted": [],
+    }
     expected_str = join_and_normalize(expected_list)
     actual_str = join_and_normalize(resp_body)
     assert actual_str == expected_str
@@ -723,9 +934,12 @@ async def test_add_flush_to_disk_to_script(ofrak_client: TestClient, firmware_zi
         "from ofrak.core import *",
         "",
         "",
-        "async def main(ofrak_context: OFRAKContext):",
+        "async def main(ofrak_context: OFRAKContext, root_resource: Optional[Resource] = None):",
         "",
-        '    root_resource = await ofrak_context.create_root_resource_from_file("firmware_zip")',
+        "    if root_resource is None:",
+        "        root_resource = await ofrak_context.create_root_resource_from_file(",
+        '            "firmware_zip"',
+        "        )",
         "",
         "    await root_resource.unpack()",
         "",